chore: Fix delete commands if entity is not found

+++ BlazorHero.CleanArchitecture.Application/Features/Products/Commands/Delete/DeleteProductCommand.cs
@@ -27,9 +27,16 @@ namespace BlazorHero.CleanArchitecture.Application.Features.Products.Commands.De
         public async Task<Result<int>> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
         {
             var product = await _unitOfWork.Repository<Product>().GetByIdAsync(command.Id);
-            await _unitOfWork.Repository<Product>().DeleteAsync(product);
-            await _unitOfWork.Commit(cancellationToken);
-            return await Result<int>.SuccessAsync(product.Id, _localizer["Product Deleted"]);
+            if (product != null)
+            {
+                await _unitOfWork.Repository<Product>().DeleteAsync(product);
+                await _unitOfWork.Commit(cancellationToken);
+                return await Result<int>.SuccessAsync(product.Id, _localizer["Product Deleted"]);
+            }
+            else
+            {
+                return await Result<int>.FailAsync(_localizer["Product Not Found!"]);
+            }
         }
     }
 }
